app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-cloud-storage
google-cloud-pubsub
requests
jinja2
python-multipart